Migrated to async/await
Migrated all `mongo` methods to their `*Async` counterparts. Migrated from `http` to `fetch`. Then, as part of this migration, we had to implement separate `AccountsTemplates.oauthServices` on client and server, with the latter being now async. Also, we made this method usable again, as it was relying on the long deprecated `Accounts.loginServiceConfiguration`, which we now replaced with `ServiceConfiguration.configurations`